% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms.r
\name{cso_storm}
\alias{cso_storm}
\title{Detect Storms According to ASM CSO Program Business Rules}
\usage{
cso_storm(x, time, threshold = 0.1, iet = 10)
}
\arguments{
\item{x}{a vector of rain amounts}

\item{time}{a vector of time stamps, used to check order if present}

\item{threshold}{a scalar representing the threshold for forward and backward sums}

\item{iet}{a scalar representing the minimum inter-event time duration}
}
\description{
The storm detection algorithm is documented in OneNote, a pdf document, and
a memo. The algorithm relies on start points that occur when it is currently
raining, a forward looking rolling sum is over the threshold, and the previous
time points within the IET (i.e., (i-iet+1):(i-1) time points where the current
point is i. A storm continues if the previous point is a storm, a backward sum
is not
}
