% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms.r
\name{decluster}
\alias{decluster}
\title{Accepts a boolean x, and a size argument. Looks for runs of true in boolean
and}
\usage{
decluster(x, min.diff)
}
\description{
Accepts a boolean x, and a size argument. Looks for runs of true in boolean
and
}
