% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseUTCm8Time}
\alias{parseUTCm8Time}
\title{Convert times expressed as UTC-08 into local times}
\usage{
parseUTCm8Time(x, ...)
}
\arguments{
\item{x}{a date, either character or POSIX}
}
\description{
The \link{package:odbc} package automatically reads dates from databases into
POSIXct objects. These must be converted into local time to allow for accurate
calculations in R.
}
