% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.aqts.R
\name{read.aqts.locations}
\alias{read.aqts.locations}
\title{Read location data from Aquarius}
\usage{
read.aqts.locations(..., full_metadata = FALSE)
}
\arguments{
\item{LocationName}{Filter results to the given location name}

\item{LocationIdentifier}{Filter results to given location identifier}

\item{TagKeys}{Filter results to locations matching all tags by key.}
}
\description{
This function uses the \code{\link[raquarius:location-requests]{raquarius::GetLocationDescriptionList()}} function to retrieve location information from the AQTS Publish Service. Setting \code{full_metadata = TRUE} will return an \link[sf:sf]{sf::st_sf} object which will allow easy plotting on a map or spatial filtering.
}
\examples{
\dontrun{
# Get a list of all locations
read.aqts.locations()
# Get all locations matching a tag
loc_list <- read.aqts.locations(TagKeys = "PAWMAP")
# Get all metadata for those locations - somewhat slower as each location is a separate API request.
read.aqts.locations(TagKeys = "PAWMAP", full_metadata = TRUE)
}

}
