% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.aqts.R
\name{read.aqts.ts}
\alias{read.aqts.ts}
\title{Read Time Series Information from Aquarius Time Series Publish Service}
\usage{
read.aqts.ts(...)
}
\arguments{
\item{LocationIdentifier}{A monitoring location identifier to filter by}

\item{Parameter}{A parameter to filter by}
}
\description{
\code{read.aqts.ts} is used to retrieve time series metadata from the AQTS Publish Service. Arguments to this function are passed unmodified to \code{\link[raquarius:ts-requests]{raquarius::GetTimeSeriesUniqueIdList()}} to retrieve matching time series. Then, metadata is retrieved with \code{\link[raquarius:ts-requests]{raquarius::GetTimeSeriesDescriptionListByUniqueId()}}.
}
\examples{
\dontrun{
ret <- read.aqts.ts(
 LocationIdentifier = c("VNB", "21B"),
 Parameter = "Temperature",
 QueryFrom = Sys.date() - 365
)
head(ret)
ret <- ret |> unnest(points)
head(ret)
}
}
