% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.aqts.R
\name{read.aqts.tsdata}
\alias{read.aqts.tsdata}
\title{Read Time Series data from Aquarius Time Series Publish Service}
\usage{
read.aqts.tsdata(..., .time_align = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{\link[raquarius:ts-requests]{raquarius::GetTimeSeriesData()}} if \code{.time_align = TRUE} or \link[raquarius:ts-requests]{raquarius::GetTimeSeriesCorrectedData} if \code{.time_align = FALSE}. Note, the arguments}

\item{.time_align}{If \code{TRUE}, the function will return multiple time aligned time series. If \code{FALSE}, will return one time series at a time with timestamps matching the original logged values for each series - which may not be time aligned.}
}
\value{
A \code{data.frame} with time series data.
}
\description{
\code{read.aqts.data} is used to retrieve time series data from the AQTS Publish Service. Arguments to this function are passed unmodified to \code{\link[raquarius:ts-requests]{raquarius::GetTimeSeriesData()}} if \code{.time_align = TRUE} or \link[raquarius:ts-requests]{raquarius::GetTimeSeriesCorrectedData} if \code{.time_align = FALSE}. Note, each of those functions takes slightly different argument values.
}
