% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bird.r
\name{read.bird}
\alias{read.bird}
\alias{read.bird.survey}
\title{Read point count data from Watershed database}
\usage{
read.bird(..., start = NULL, end = NULL, server = NULL)

read.bird.survey(
  ...,
  start = NULL,
  end = NULL,
  server = NULL,
  date.field = "survey_start"
)
}
\arguments{
\item{...}{query parameters in \code{tag = value} form}

\item{start}{an object of class Date; the earliest survey start date to retrieve}

\item{end}{an object of class Date; the latest survey start date to retrieve}

\item{server}{Alternate server for the WATERSHED database - for access to production or test instances.}
}
\description{
Reads point count data and survey information from BES watershed database
}
\section{Functions}{
\itemize{
\item \code{read.bird}: Read point count data

\item \code{read.bird.survey}: Read point count survey information
}}

\examples{
\dontrun{
#Get some data
d <- read.bird(start = as.Date('2017-01-01'))
}
}
