% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.esri.R
\name{read.esri}
\alias{read.esri}
\title{Read from spatial data from City of Portland spatial databases}
\usage{
read.esri(database = "EGH_PUBLIC", layer, query = NULL, ...)
}
\arguments{
\item{database}{Either EGH_PUBLIC on GISDB1 or SWSP on BESDBPROD1}

\item{layer}{a fully qualified SQL table name, `[schema].[table]`}

\item{query}{a query to submit directly to server; `Shape.STAsBinary() as geometry` is necessary in select statement}

\item{...}{additional arguments passed to st_read. The correct CRS, usually 2913, will need to be passed.}
}
\value{
an `sf` object, additionally inheriting `tbl_df`, `tbl`
}
\description{
Reads from a defined list of databases.
}
