% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.flow.R
\name{read.flow}
\alias{read.flow}
\title{Retrieve flow data from Janus}
\usage{
read.flow(..., start = NULL, end = NULL, dsn = "JANUS_REP_64",
  date.field = "reading_datetime")
}
\arguments{
\item{...}{query arguments in the form 'field = value'}

\item{start}{start date of data range}

\item{end}{end date of data range}

\item{dsn}{the dsn for the JANUS database}

\item{date.field}{the field used for the start and end query arguments. The default value is "reading_datetime"}
}
\value{
a data frame of results with the fields: \cr
project_name\cr
location_code\cr
location_description\cr
manhole_hansen_id \cr
reading_datetime \cr
depth_inches \cr
velocity_fps \cr
depth_qualifier \cr
velocity_qualifier \cr
flow_cfs_AxV \cr
}
\description{
Submits a query to a table join allowing for queries by Hansen ID and Date
}
\details{
To query by Hansen ID use the field name mv.manhole_hansen_id. To query a specific
date range use start and end arguments, otherwise all available data at that location
will be returned.

The user running R must have permission to access the database view and must have
an ODBC Data Source set up for the Janus database on the current computer.
}
\author{
Peter Bryant peter.bryant@portlandoregon.gov
}
