% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.flow.R
\name{read.janus.flow}
\alias{read.janus.flow}
\title{Retrieve flow data from Janus}
\usage{
read.janus.flow(
  ...,
  start = NULL,
  end = NULL,
  server = NULL,
  date.field = "reading_datetime"
)
}
\arguments{
\item{...}{query arguments in the form 'field = value'}

\item{start}{start date of data range}

\item{end}{end date of data range}

\item{server}{Alternate server for the JANUS database - for access to production or test instances.}

\item{date.field}{the field used for the start and end query arguments. The default value is "reading_datetime"}
}
\value{
a data frame of results with the fields: \cr
project_name\cr
location_code\cr
location_description\cr
manhole_hansen_id \cr
reading_datetime \cr
depth_inches \cr
velocity_fps \cr
depth_qualifier \cr
velocity_qualifier \cr
flow_cfs_AxV \cr
}
\description{
Submits a query to a table join allowing for queries by Hansen ID and date.
}
\details{
To query by Hansen ID use the field name mv.manhole_hansen_id. To query a specific date range use start and end arguments, otherwise all available data at that location will be returned.

The user running R must have permission to access the view (V_METER_VISIT_LIST_NARROW) and table (METER_READING) used in the query. The user must have an ODBC Data Source set up for the Janus database on the current computer. The reporting server version of the database is currently BESREPORTS.JANUS.
}
\author{
Peter Bryant \email{peter.bryant@portlandoregon.gov}
}
