% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rain.r
\name{read.rain}
\alias{read.rain}
\title{Retrieve rain data from City of Portland Hydra network.}
\usage{
read.rain(station = 160, start = end - 7, end = Sys.Date(),
  daypart = c("day", "hour", "minute", "month", "year"), interval = 1,
  dsn = NULL, format = T)
}
\arguments{
\item{station}{a vector of hydra station codes; the default is the WPCL gauge}

\item{start}{the start date as a \code{Date} object}

\item{end}{the end date as a \code{Date} object}

\item{daypart}{the unit in which the interval is specified}

\item{interval}{the interval over which the data should be summarized}

\item{dsn}{Alternate dsn for the NEPTUNE database - for access to production or test instances.}

\item{format}{logical, TRUE will perform some data formatting, FALSE will return the data.frame exactly
as it was returned by the database}
}
\value{
a data.frame of rain data
}
\description{
The City of Portland operates a rain gauge network.  This function retrieves
rain data from the SQL server database using the USP_MODEL_RAIN stored procedure.
The stored procedure summarizes the data within a prespecified period of time
(e.g., hourly) and returns the summarized data rather than raw tips.
}
\details{
A DSN for the Hydra database must be set up on your machine. The current location of
the stored procedure is BESDBPROD2.NEPTUNE.USP_MODEL_RAIN.  The name of
that DSN is then passed to the function via the \code{DSN} argument.
}
