% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms.r
\name{storm_delineation}
\alias{storm_delineation}
\alias{cso_storm}
\alias{pot_storm}
\title{Storm delineation methods}
\usage{
cso_storm(x, time, threshold = 0.1, iet = 10)

pot_storm(x, time, threshold, iet)
}
\arguments{
\item{x}{a vector of rain amounts}

\item{time}{a vector of time stamps, used to check order if present. If this
argument is not used, the vector of rain amounts \code{x} is assumed to be
ordered}

\item{threshold}{a scalar representing the method threshold - this parameter
has different interpretations for each delineation method}

\item{iet}{a scalar representing the minimum inter-event time duration - the
parameter has essentially the same meaning across methods, but the restriction
is enforced in very different ways and so the results will differ across methods}
}
\description{
These two functions assign determine whether each time point in a time series
of rainfall totals is in a storm event or not. The \code{cso_storm} function
implements an algorithm that has been used to determine storm events for CSO
program reporting and analysis. The \code{pot_storm} method implements a more
general purpose peak over threshold method along with a minimum inter-event
time. Any rain value over the threshold is part of a storm and time stamps less
than \code{iet} distance from each other are considered part of the same storm.
}
\details{
The \code{cso_stor} storm detection algorithm is documented in
OneNote and a pdf document maintained by the ASM team. Please contact them for
access to the most recent version. Jason Law has also documented the algorithm
in a memo with more detailed mathematical description of the algorithm.
The algorithm relies on start points that
occur when it is currently raining, a forward looking rolling sum is over
the threshold, and the previous time points within the IET
(i.e., (i-iet+1):(i-1) time points where the current point is i. A storm
continues if the previous value is a storm and the current value is not a
stopping point. The current value is a stopping point if, a backward sum is not over the
threshold and the forward
}
