% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms.r
\name{summarize_delineation}
\alias{summarize_delineation}
\title{Summarize storm delineation for entire period}
\usage{
summarize_delineation(storms)
}
\arguments{
\item{storms}{A data.frame created using \link{summarize_storms}}
}
\value{
a single row data.frame with columns summarizing the performance of
the delineation
}
\description{
This function summarizes the output of \code{summarize_storms}. It calculates
summaries like the total rainfall over the period, the amount contained in
the storms, the proportion of storm rainfall out of total rainfall,
total duration of storms, average intensity of storms extracted.
}
\details{
This function can be used to summarize the performance of the storm
delineation methods (e.g., \link{pot_storm} or \link{cso_storm}) or different
parameter values passed to these methods
}
