% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.janus.r
\name{read.janus}
\alias{read.janus}
\title{Submits a query to JANUS database}
\usage{
read.janus(
  ...,
  start = NULL,
  end = NULL,
  what = c("data", "location"),
  server = NULL,
  date.field = "sample_end_time",
  as.text = T,
  simple = F
)
}
\arguments{
\item{...}{query arguments in the form 'field = value'.}

\item{start}{start date of data range}

\item{end}{end date of data range}

\item{what}{whether to return data or locations only. \code{start} and \code{end} are ignored if locations}

\item{date.field}{the field used for the start and end query arguments.}

\item{as.text}{logical, return a text (TRUE) or numeric (FALSE) nondetect indicator in column `nd`.}

\item{dsn}{the dsn for the JANUS database on the current system}
}
\value{
a dataframe of results
}
\description{
Submits a query to JANUS database view JANUS_ELEMENT using
arbitrary tag=value query parameters.
}
\details{
The user running R must have permission to access the database view and
must have a ODBC Data Source set up for the view on the current computer.
The reporting server version of the table is currently BESREPORTS.WATERSHED.JANUS_ELEMENT.
In addition, different ODBC connections must be set up for 32 and 64 bit
versions of R. The \code{RODBC::odbcDataSources()} function can be used to
check the currently available data sources from within R.
}
\examples{
\dontrun{
#List the available fields
library(RODBC)
con <- BESdata:::dbConnect("JANUS")
sqlColumns(con, 'JANUS_ELEMENT')
#Get the last 2 weeks of data
read.janus(start = Sys.Date() - 14)
}
}
\author{
Jason Law \email{jason.law@portlandoregon.gov}
}
